requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
